+++ app/Http/Controllers/Admin/StoreController.php
@@ -9,9 +9,11 @@ use App\Http\Resources\Admin\StoreResource;
 use App\Models\Admin;
 use App\Models\Area;
 use App\Models\Company;
+use App\Models\Expenses;
 use App\Models\Product;
 use App\Models\Receipt;
 use App\Models\ReceiptProduct;
+use App\Models\Revenu;
 use App\Models\Store;
 use App\Models\StoreProduct;
 use App\Models\Treasury;
@@ -29,12 +31,23 @@ class StoreController extends Controller
      */
     public function index()
     {
-        $stores = Store::with('keeper:id,name,store_id','finance_manager:id,name,store_id','area')->paginate(10);
+        if(!in_array(91,permissions())){
+            abort(403);
+        }
+        if(auth()->user()->type == 'المسؤول العام'){
+            $stores = Store::paginate(10);
+        }else{
+            $stores = Store::where('id',auth()->user()->store_id)->first();
+        }
         return view('admin.store.index',compact('stores'));
     }
 
     public function getStores(){
-        $stores = Store::with('keeper:id,name,store_id','finance_manager:id,name,store_id','area');
+        if(auth()->user()->type == 'admin'){
+            $stores = Store::get();
+        }else{
+            $stores = Store::where('id',auth()->user()->store_id);
+        }
         return datatables($stores)->make(true);
     }
 
@@ -45,13 +58,19 @@ class StoreController extends Controller
      */
     public function create()
     {
+        if(!in_array(90,permissions())){
+            abort(403);
+        }
         $storekeepers = Admin::where('type','امين مخزن')->where('store_id',null)->get();
         $procurement_officers = Admin::where('type','مسؤول المشتريات')->where('store_id',null)->get();
         $finance_officers = Admin::where('type','مسؤول الماليه')->where('store_id',null)->get();
         $salesmen = Admin::where('type','بائع')->where('store_id',null)->get();
-        $areas = Area::where('status','تفعيل')->get();
+        $areas = Area::where('status','تفعيل')->where('store_id',null)->get();
         $status = Store::getEnumValues('stores','status');
+
         return view('admin.store.create',compact('storekeepers','procurement_officers','finance_officers','salesmen','areas','status'));
+
+
     }
 
     /**
@@ -64,14 +83,17 @@ class StoreController extends Controller
     {
         $store = Store::create([
             'name' => $request->name,
-            'area_id' => $request->area_id,
-            'store_keeper_id' => $request->storekeeper_id,
-            'store_finance_manager_id' => $request->finance_officer_id,
             'address' => $request->address,
             'longitude' => $request->longitude,
             'latitude' => $request->latitude,
             'status' => $request->status,
         ]);
+
+        Treasury::create([
+            'store_id' => $store->id,
+            'price' => 0,
+        ]);
+        Area::whereIn('id',$request->area_id)->update(['store_id'=>$store->id]);
         Admin::whereIn('id',$request->finance_officers)->update(['store_id'=>$store->id]);
         Admin::whereIn('id',$request->storekeepers)->update(['store_id'=>$store->id]);
         Treasury::create(['store_id'=>$store->id]);
@@ -79,6 +101,9 @@ class StoreController extends Controller
     }
 
     public function getStoreProducts($store_id){
+        if(!in_array(30,permissions())){
+            abort(403);
+        }
         $products = Product::whereHas('stores',function($q) use($store_id){
                         $q->where('stores_products.store_id',$store_id)->select('stores.id');
                     })
@@ -86,8 +111,15 @@ class StoreController extends Controller
                             $q->where('stores_products.store_id',$store_id)->select('stores.id');
                         },'company:id,name','category:id,name'])
                     ->select('products.name','products.id','wholesale_type','item_type','company_id','category_id')->paginate(10);
-                  $products = StoreResource::collection($products);
-        return view('admin.store.product',compact('products','store_id'));
+        $products = StoreResource::collection($products);
+
+        $treasury = Treasury::where('store_id',$store_id)->sum('price');
+        $revenues = Revenu::where('store_id',$store_id)->get()->Sum('price');
+        $expenses = Expenses::where('store_id',$store_id)->get()->Sum('price');
+
+        return view('admin.store.product',compact('products','store_id','treasury','revenues','expenses'));
+
+
     }
 
     public function getStoreProductsTable($store_id){
@@ -104,6 +136,9 @@ class StoreController extends Controller
     }
 
     public function editStoreProduct($product_id,$store_id){
+        if(!in_array(77,permissions())){
+            abort(403);
+        }
         $product = Product::with(['stores'=>function($q) use($store_id){
             $q->where('stores_products.store_id',$store_id)->select('stores.id')->first();
         }])->where('products.id',$product_id)
@@ -169,9 +204,16 @@ class StoreController extends Controller
             $q->where('store_id',$store->id)->orWhere('store_id',null);
         })->get();
 
-        $areas = Area::where('status','تفعيل')->get();
+        $areas = Area::where('status','تفعيل')->where(function($q) use($store){
+            $q->where('store_id',null)->orWhere('store_id',$store->id);
+        })->get();
+        $store_areas = Area::where('status','تفعيل')->where('store_id',$store->id)->pluck('id')->toArray();
         $status = Store::getEnumValues('stores','status');
-        return view('admin.store.edit',compact('keepers','storekeepers','finance_officers','store_finance_officers','salesmen','stor_salesmen','areas','status','store'));
+        if(in_array(92,permissions())){
+            return view('admin.store.edit',compact('keepers','storekeepers','finance_officers','store_finance_officers','store_areas','salesmen','stor_salesmen','areas','status','store'));
+        }else{
+            abort(403);
+        }
     }
 
     /**
@@ -185,9 +227,6 @@ class StoreController extends Controller
     {
         $store->update([
             'name' => $request->name,
-            'area_id' => $request->area_id,
-            'store_keeper_id' => $request->storekeeper_id,
-            'store_finance_manager_id' => $request->finance_officer_id,
             'address' => $request->address,
             'longitude' => $request->longitude,
             'latitude' => $request->latitude,
@@ -195,6 +234,9 @@ class StoreController extends Controller
         ]);
 
         Admin::where('store_id',$store->id)->update(['store_id'=>null]);
+        Area::where('store_id',$store->id)->update(['store_id'=>null]);
+
+        Area::whereIn('id',$request->area_id)->update(['store_id'=>$store->id]);
 
         Admin::whereIn('id',$request->finance_officers)->update(['store_id'=>$store->id]);
         Admin::whereIn('id',$request->storekeepers)->update(['store_id'=>$store->id]);
@@ -210,9 +252,13 @@ class StoreController extends Controller
      */
     public function destroy(Store $store)
     {
+        if(!in_array(93,permissions())){
+            abort(403);
+        }
         $recepits =Receipt::where('store_id',$store->id)->pluck('id');
         ReceiptProduct::whereIn('receipt_id', $recepits)->delete();
         Receipt::where('store_id',$store->id)->delete();
+        Treasury::where('store_id',$store->id)->delete();
         StoreProduct::where('store_id',$store->id)->delete();
         $store->delete();
     }
